Rename test files for clarity
