+++ tests/testStroke.ts
